pyproject.toml
[project]
name = "calendar-reminder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "google-api-python-client>=2.175.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "loguru>=0.7.3",
    "uvicorn>=0.35.0",
]
